move noticias components and containers into noticias directories. move uploads and upload field into fields directory
